+++ pdfminer/image.py
@@ -37,7 +37,10 @@ class ImageWriter:
         """Save an LTImage to disk"""
         filters = image.stream.get_filters()
 
-        if filters[-1][0] in LITERALS_DCT_DECODE:
+        if not filters:
+            name = self._save_bytes(image)
+
+        elif filters[-1][0] in LITERALS_DCT_DECODE:
             name = self._save_jpeg(image)
 
         elif filters[-1][0] in LITERALS_JPX_DECODE: