+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Extract images with Indexed colorspace correctly
 - Save CMYK JPEG images as CMYK
 - Save other CMYK images in .tiff format
+- Don't throw when saving image with no filters
 
 ## [20250506]